Copy files from v1 and update import path to v2
Add type alias and rename based struct for List

Regenerate *_list.go

Generate resource files (wip)

Remove Base struct + regenerate request files

Prepend omise package prefix to field type + auto import package

Generate nested resources (wip)

Add resource prefix for list actions

Fix duplicate refund struct types

Append List struct to list operations

Update ordering path to v2

Generate extra params for endpoints

Regenerate operations/*

Generate all params

Generate Vault endpoint for token

Update test to include v2

Update fixtures path

Generate comments

Ignore golint for now (focus on tests)

Generate omitempty tags

Generate struct pointers

Reorder json fields in tests

Regenerate int64 and int types

Update fixtures path

Force non-null fields

Generate *bool for charge.capture

Generate documentation (wip)

Generate documentation (wip)

Add tests for capability

Generate float64 for floating point fields

Add tests for forex

Add DateString type for date fields in params

Generate custom DateString type

Fix test failed due to case-sensitivity in linux

Update tests for schedule + make Weekday non-pointer

Fix more tests in schedule resource

Make Schedule's start_date and end_date nullable

Generate ListChargeSchedules struct

Re-generate ListChargeSchedules struct

Make enum types non-reference

Update tests and fixtures for source

Add tests for system_info (/)

Generate search types

Re-generate event.data

Update tests for schedule

Add network tests for source

Re-enable golint for v2

Add comments for enums

Generate comments for enums

Generate more comments

Remove old codes

Regenerate files with newlines at the end of files

Regenerate files with newlines at the end of files (cont)

Regenerate files with newlines at the end of files (cont)

Regenerate files with newlines at the end of files (done)

Cleanup go mod dependencies

+++ v2/balance.go
@@ -6,7 +6,8 @@ type Balance struct {
 	Base
 	Currency string `json:"currency"`
 	Location string `json:"location"`
-	Reserve int `json:"reserve"`
-	Total int `json:"total"`
-	Transferable int `json:"transferable"`
+	Reserve int64 `json:"reserve"`
+	Total int64 `json:"total"`
+	Transferable int64 `json:"transferable"`
 }
+
