+++ v2/payment_method.go
@@ -1,9 +1,11 @@
 package omise
 
+// PaymentMethod represents resource object.
 type PaymentMethod struct {
 	Base
 	CardBrands []string `json:"card_brands"`
 	Currencies []string `json:"currencies"`
-	InstallmentTerms []int `json:"installment_terms"`
+	InstallmentTerms []int64 `json:"installment_terms"`
 	Name string `json:"name"`
 }
+
