+++ v2/payment_reference.go
@@ -4,12 +4,13 @@ import (
 	"time"
 )
 
+// PaymentReference represents resource object.
 type PaymentReference struct {
 	Base
 	Barcode string `json:"barcode"`
-	CustomerAmount int `json:"customer_amount"`
+	CustomerAmount int64 `json:"customer_amount"`
 	CustomerCurrency string `json:"customer_currency"`
-	CustomerExchangeRate int `json:"customer_exchange_rate"`
+	CustomerExchangeRate float64 `json:"customer_exchange_rate"`
 	DeviceID string `json:"device_id"`
 	ExpiresAt time.Time `json:"expires_at"`
 	OmiseTaxID string `json:"omise_tax_id"`
@@ -18,3 +19,4 @@ type PaymentReference struct {
 	ReferenceNumber2 string `json:"reference_number_2"`
 	VaCode string `json:"va_code"`
 }
+
