+++ v2/platform_fee.go
@@ -1,8 +1,10 @@
 package omise
 
+// PlatformFee represents resource object.
 type PlatformFee struct {
 	Base
-	Amount int `json:"amount"`
-	Fixed int `json:"fixed"`
-	Percentage int `json:"percentage"`
+	Amount int64 `json:"amount"`
+	Fixed int64 `json:"fixed"`
+	Percentage float64 `json:"percentage"`
 }
+
