+++ v2/transfer_scheduling.go
@@ -4,8 +4,9 @@ package omise
 // See https://www.omise.co/transfer-schedules-api for more information.
 type TransferScheduling struct {
 	Base
-	Amount *int `json:"amount"`
+	Amount *int64 `json:"amount"`
 	Currency string `json:"currency"`
-	PercentageOfBalance *int `json:"percentage_of_balance"`
+	PercentageOfBalance *float64 `json:"percentage_of_balance"`
 	Recipient string `json:"recipient"`
 }
+
